requirements.txt
python-dotenv==1.0.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.82.0
functions-framework==3.3.0
requests==2.31.0
pytz==2023.3
